refactor: move config migrations to shared
