+++ .changeset/shared-config-migrations.md
@@ -1,0 +1,5 @@
+---
+"@aliou/pi-guardrails": patch
+---
+
+Move config migrations into shared modules and only show onboarding when no guardrails config exists.
