+++ extensions/guardrails/utils/warnings.ts
@@ -1,7 +1,5 @@
-/**
- * Module-level warnings queue for messages that arise before any session
- * context is available (config loading, migration, pattern compilation).
- *
- * Drained and reported via ctx.ui.notify in the session_start handler.
- */
-export const pendingWarnings: string[] = [];
+export {
+  addPendingWarning,
+  drainPendingWarnings,
+  getPendingWarnings,
+} from "../../../src/shared/warnings";
