+++ package.json
@@ -31,7 +31,7 @@
     "README.md"
   ],
   "dependencies": {
-    "@aliou/pi-utils-settings": "^0.11.2",
+    "@aliou/pi-utils-settings": "^0.13.0",
     "@aliou/sh": "^0.1.0"
   },
   "peerDependencies": {
