+++ src/shared/config/index.ts
@@ -1,0 +1,18 @@
+export { DEFAULT_CONFIG } from "./defaults";
+export { configLoader } from "./loader";
+export {
+  CURRENT_VERSION,
+  globalConfigMigrations,
+  migrations,
+  normalizeAllowedPaths,
+} from "./migration";
+export type {
+  DangerousPattern,
+  GuardrailsConfig,
+  PathAccessConfig,
+  PathAccessMode,
+  PatternConfig,
+  PolicyRule,
+  Protection,
+  ResolvedConfig,
+} from "./types";
