+++ src/shared/config/migration/003-strip-command-explainer-fields.ts
@@ -1,0 +1,42 @@
+import { addPendingWarning } from "../../warnings";
+import type { GuardrailsConfig } from "../types";
+import { CURRENT_VERSION } from "./version";
+
+const REMOVED_PERMISSION_GATE_KEYS = [
+  "explainCommands",
+  "explainModel",
+  "explainTimeout",
+] as const;
+
+export function shouldRun(config: GuardrailsConfig): boolean {
+  const raw = config as Record<string, unknown>;
+  const permissionGate = raw.permissionGate as
+    | Record<string, unknown>
+    | undefined;
+  if (!permissionGate) return false;
+
+  for (const key of REMOVED_PERMISSION_GATE_KEYS) {
+    if (key in permissionGate) return true;
+  }
+
+  return false;
+}
+
+export function run(config: GuardrailsConfig): GuardrailsConfig {
+  addPendingWarning(
+    "[guardrails] permissionGate.explainCommands, explainModel, and explainTimeout " +
+      "have been removed. These fields will be stripped from your config.",
+  );
+
+  const cleaned = structuredClone(config) as Record<string, unknown>;
+  const permissionGate = cleaned.permissionGate as
+    | Record<string, unknown>
+    | undefined;
+  if (permissionGate) {
+    for (const key of REMOVED_PERMISSION_GATE_KEYS) {
+      delete permissionGate[key];
+    }
+  }
+  cleaned.version = CURRENT_VERSION;
+  return cleaned as GuardrailsConfig;
+}
