+++ src/shared/config/migration/version.ts
@@ -1,0 +1,7 @@
+/**
+ * Config schema version.
+ *
+ * Keep this independent from package.json version.
+ * Bump only when config schema/default migration markers change.
+ */
+export const CURRENT_VERSION = "0.9.0-20260327";
